Move loadSchemaFromPath to utils file
