+++ packages/app/src/cli/models/extensions/specifications/flow_action.ts
@@ -7,8 +7,7 @@ import {
   validateReturnTypeConfig,
 } from '../../../services/flow/validation.js'
 import {serializeFields} from '../../../services/flow/serialize-fields.js'
-import {joinPath} from '@shopify/cli-kit/node/path'
-import {glob, readFile} from '@shopify/cli-kit/node/fs'
+import {loadSchemaPatchFromPath} from '../../../services/flow/utils.js'
 import {zod} from '@shopify/cli-kit/node/schema'
 
 export const FlowActionExtensionSchema = BaseSchemaWithHandle.extend({
@@ -33,25 +32,6 @@ export const FlowActionExtensionSchema = BaseSchemaWithHandle.extend({
 
   return configurationPageIsValid && settingsFieldsAreValid && returnTypeIsValid
 })
-
-/**
- * Loads the schema from the partner defined file.
- */
-export const loadSchemaPatchFromPath = async (extensionPath: string, patchPath: string | undefined) => {
-  if (!patchPath) {
-    return ''
-  }
-
-  const path = await glob(joinPath(extensionPath, patchPath))
-
-  if (path.length > 1) {
-    throw new Error('Multiple files found for schema patch path')
-  } else if (path.length === 0) {
-    throw new Error('No file found for schema patch path')
-  }
-
-  return readFile(path[0] as string)
-}
 
 /**
  * Extension specification with all properties and methods needed to load a Flow Action.
